add BUTTERAUGLI='n' to disable butteraugli install by default

+++ optimise-images.sh
@@ -44,7 +44,7 @@
 # http://dinbror.dk/blog/blazy/?ref=demo-page
 ########################################################################
 DT=$(date +"%d%m%y-%H%M%S")
-VER='5.0'
+VER='5.1'
 DEBUG='n'
 
 # Used for optimise-age mod, set FIND_IMGAGE in minutes. So to only
@@ -203,6 +203,7 @@ LOGNAME_PROFILE="profile-log-${DT}.log"
 LOG_PROFILE="${LOGDIR}/${LOGNAME_PROFILE}"
 BENCHDIR='/home/optimise-benchmarks'
 
+BUTTERAUGLI='n'
 GUETZLI_BIN='/opt/guetzli/bin/Release/guetzli'
 BUTTERAUGLI_BIN='/usr/bin/butteraugli'
 GM_BIN='/usr/bin/gm'
@@ -243,7 +244,7 @@ if [[ ! -f /etc/centminmod-release && "$CENTOS_SEVEN" = '7' && ! "$(rpm -qa Imag
   rpm -Uvh remi-release-7.rpm
   echo
   echo "Install YUM dependencies"
-  yum -q -y install ImageMagick6 ImageMagick6-devel ImageMagick6-c++ ImageMagick6-c++-devel ImageMagick6-libs LibRaw libpng-devel wget bc git make nasm gcc gcc-c++ coreutils optipng jpegoptim jpegtran GraphicsMagick sysstat util-linux --enablerepo=remi
+  yum -q -y install ImageMagick6 ImageMagick6-devel ImageMagick6-c++ ImageMagick6-c++-devel ImageMagick6-libs LibRaw libjpeg-turbo-devel libpng-devel wget bc git make nasm gcc gcc-c++ coreutils optipng jpegoptim jpegtran GraphicsMagick sysstat util-linux --enablerepo=remi
   echo "YUM dependencies installed"
 elif [[ ! -f /etc/centminmod-release && "$CENTOS_SIX" = '6' && ! "$(rpm -qa ImageMagick6 | grep -o 'ImageMagick6')" ]]; then
   echo
@@ -252,7 +253,7 @@ elif [[ ! -f /etc/centminmod-release && "$CENTOS_SIX" = '6' && ! "$(rpm -qa Imag
   rpm -Uvh remi-release-6.rpm
   echo
   echo "Install YUM dependencies"
-  yum -q -y install ImageMagick6 ImageMagick6-devel ImageMagick6-c++ ImageMagick6-c++-devel ImageMagick6-libs LibRaw libpng-devel wget bc git make nasm gcc gcc-c++ coreutils optipng jpegoptim jpegtran GraphicsMagick sysstat util-linux-ng --enablerepo=remi
+  yum -q -y install ImageMagick6 ImageMagick6-devel ImageMagick6-c++ ImageMagick6-c++-devel ImageMagick6-libs LibRaw libjpeg-turbo-devel libpng-devel wget bc git make nasm gcc gcc-c++ coreutils optipng jpegoptim jpegtran GraphicsMagick sysstat util-linux-ng --enablerepo=remi
   echo "YUM dependencies installed"
 fi
 
@@ -477,16 +478,24 @@ mozjpeg_install() {
 }
 
 butteraugli_install() {
-  echo
-  echo "installing butteraugli"
-  cd /opt
-  rm -rf butteraugli
-  git clone https://github.com/google/butteraugli
-  cd butteraugli/butteraugli
-  make
-  \cp -af butteraugli /usr/bin/butteraugli
-  BUTTERAUGLI_BIN='/usr/bin/butteraugli'
-  echo "installed butteraugli" 
+  if [[ "$BUTTERAUGLI" = [yY] ]]; then
+    echo
+    echo "installing butteraugli"
+    cd /opt
+    rm -rf butteraugli
+    git clone https://github.com/google/butteraugli
+    cd butteraugli/butteraugli
+    make
+    if [ -f butteraugli ]; then
+      \cp -af butteraugli /usr/bin/butteraugli
+    fi
+    if [ -f /usr/bin/butteraugli ]; then
+      BUTTERAUGLI_BIN='/usr/bin/butteraugli'
+      echo "installed butteraugli"
+    else
+      echo "failed to install butteraugli"
+    fi
+  fi
 }
 
 guetzli_install() {
